"Use geolocation to get user's current position and send it to the server"
This commit updates the `App.js` file to use the `navigator.geolocation.getCurrentPosition()` method to get the user's current position, and then sends this information to the server using the `axios.post()` method. Previously, the `axios.get()` method was used to send a request to the server without any data. The `main()` function has been updated to accept a `body` parameter, which contains the user's latitude and longitude. These values are then sent to the server in the request body.